fastapi-app/requirements.txt
aiopg==1.4.0
anyio==3.6.2
appnope==0.1.3
asttokens==2.1.0
async-timeout==4.0.2
asyncpg==0.27.0
asyncpg-listen==0.0.6
backcall==0.2.0
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.4
databases==0.6.1
decorator==5.1.1
Deprecated==1.2.13
dnspython==2.2.1
docopt==0.6.2
email-validator==1.3.0
executing==1.2.0
fastapi==0.88.0
greenlet==2.0.1
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
ipython==8.7.0
jedi==0.18.2
jwcrypto==1.4.2
logging-tree==1.9
logx==0.1.1579232358
matplotlib-inline==0.1.6
numpy==1.23.5
pandas==1.5.2
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pgnotify==0.1.1561372201
pickleshare==0.7.5
pipreqs==0.4.11
prompt-toolkit==3.0.33
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
PyJWT==2.6.0
pyodbc==4.0.35
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
pytz==2022.6
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.44
stack-data==0.6.2
starlette==0.22.0
traitlets==5.5.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
watchfiles==0.18.1
wcwidth==0.2.5
websocket-client==1.4.2
websockets==10.4
wrapt==1.14.1
yarg==0.1.9
